Implement new format proposal

+++ include/xgboost/data.h
@@ -39,7 +39,7 @@ enum class DataType : uint8_t {
 class MetaInfo {
  public:
   /*! \brief number of data fields in MetaInfo */
-  static constexpr uint32_t kNumField = 7;
+  static constexpr uint64_t kNumField = 7;
 
   /*! \brief number of rows in the data */
   uint64_t num_row_{0};
